backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
pydantic==1.10.7
pyswisseph==2.10.03
